Refactor Jest scripts directory structure
Introduces a more consistent naming scheme.
