+++ scripts/jest/typescript/jest.d.ts
