+++ scripts/jest/typescript/preprocessor.js
